refactor and move get_predeployed_acc_execute_args to rpc utils
